requirements/requirements.txt
alembic==1.13.1
amqp==5.2.0
apispec==6.6.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
Authlib==1.3.1
bemserver-core==0.18.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
flask-smorest==0.43.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==0.29.0
numpy==1.26.4
oso==0.27.3
packaging==24.0
pandas==2.2.2
Pint==0.23
prompt_toolkit==3.0.46
psycopg==3.1.19
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==4.6.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.3
